+++ packages/pluggableWidgets/video-player-native/src/VideoPlayer.tsx
@@ -1,7 +1,17 @@
 import { flattenStyles } from "@mendix/piw-native-utils-internal";
-import { createElement, ReactElement, useEffect, useState } from "react";
-import { ActivityIndicator, Text, View } from "react-native";
-import Video from "react-native-video";
+import { createElement, ReactElement, useEffect, useRef, useState, Fragment } from "react";
+import {
+    ActivityIndicator,
+    Platform,
+    StatusBar,
+    Text,
+    TouchableOpacity,
+    TouchableWithoutFeedback,
+    View
+} from "react-native";
+import Video, { OnProgressData, VideoProperties } from "react-native-video";
+import Modal from "react-native-modal";
+import Icon from "react-native-vector-icons/MaterialIcons";
 import { VideoPlayerProps } from "../typings/VideoPlayerProps";
 import { defaultVideoStyle, VideoStyle } from "./ui/Styles";
 import { isAvailable } from "@mendix/piw-utils-internal";
@@ -16,8 +26,15 @@ const enum StatusEnum {
 
 export function VideoPlayer(props: VideoPlayerProps<VideoStyle>): ReactElement {
     const [styles, setStyles] = useState(flattenStyles(defaultVideoStyle, props.style));
+    const timeoutRef = useRef<NodeJS.Timeout>();
+
+    const playerRef = useRef<Video>(null);
+    const fullScreenPlayerRef = useRef<Video>(null);
     const [status, setStatus] = useState(StatusEnum.NOT_READY);
     const [videoAspectRatio, setVideoAspectRatio] = useState(0);
+    const [fullScreen, setFullScreen] = useState(false);
+    const [showControls, setShowControls] = useState(false);
+    const [currentPlayTime, setCurrentPlayTime] = useState(0);
 
     useEffect(() => {
         const alteredStyles = deepmerge({}, styles);
@@ -36,27 +53,133 @@ export function VideoPlayer(props: VideoPlayerProps<VideoStyle>): ReactElement {
         setStyles(alteredStyles);
     }, [props.style, props.aspectRatio, videoAspectRatio]);
 
+    useEffect(() => {
+        if (props.showControls) {
+            handleShowControls(true);
+        }
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, [fullScreen]);
+
+    function handleShowControls(setShown?: boolean): void {
+        if (!props.showControls) {
+            return;
+        }
+
+        if (timeoutRef.current) {
+            clearTimeout(timeoutRef.current as NodeJS.Timeout);
+        }
+
+        if (!showControls || setShown) {
+            setShowControls(true);
+            timeoutRef.current = global.setTimeout(() => {
+                setShowControls(false);
+            }, 5000);
+        } else {
+            setShowControls(false);
+        }
+    }
+
+    function fullScreenHandler(isFullScreen: boolean): void {
+        setFullScreen(isFullScreen);
+        StatusBar.setHidden(isFullScreen);
+    }
+
+    const videoProps: VideoProperties = {
+        testID: props.name,
+        source: { uri: isAvailable(props.videoUrl) ? props.videoUrl.value : undefined },
+        muted: props.muted,
+        repeat: props.loop,
+        controls: props.showControls,
+        onLoadStart: () => setStatus(StatusEnum.LOADING),
+        onError: () => {
+            setStatus(StatusEnum.ERROR);
+        },
+        useTextureView: false,
+        resizeMode: props.aspectRatio ? "contain" : "stretch",
+        onProgress: ({ currentTime }: OnProgressData) => {
+            setCurrentPlayTime(currentTime);
+        }
+    };
+
+    const isAndroid = Platform.OS === "android";
+
     return (
-        <View style={styles.container}>
-            {status === StatusEnum.LOADING && <ActivityIndicator color={styles.indicator.color} size="large" />}
-            {status === StatusEnum.ERROR && <Text style={styles.errorMessage}>The video failed to load :(</Text>}
-            <Video
-                testID={props.name}
-                source={{ uri: isAvailable(props.videoUrl) ? props.videoUrl.value : undefined }}
-                paused={!props.autoStart}
-                muted={props.muted}
-                repeat={props.loop}
-                controls={props.showControls}
-                onLoadStart={() => setStatus(StatusEnum.LOADING)}
-                onLoad={data => {
-                    setStatus(StatusEnum.READY);
-                    setVideoAspectRatio(data.naturalSize.width / data.naturalSize.height);
-                }}
-                onError={() => setStatus(StatusEnum.ERROR)}
-                style={status !== StatusEnum.READY ? { height: 0 } : styles.video}
-                useTextureView={false}
-                resizeMode={props.aspectRatio ? "contain" : "stretch"}
-            />
-        </View>
+        <Fragment>
+            {isAndroid && (
+                <Modal
+                    isVisible={fullScreen}
+                    style={styles.fullScreenVideoPlayer}
+                    onBackButtonPress={() => {
+                        fullScreenHandler(false);
+                    }}
+                >
+                    <TouchableWithoutFeedback onPress={() => handleShowControls()}>
+                        <Video
+                            {...videoProps}
+                            ref={fullScreenPlayerRef}
+                            paused={false}
+                            onLoad={data => {
+                                setStatus(StatusEnum.READY);
+                                setVideoAspectRatio(data.naturalSize.width / data.naturalSize.height);
+                                fullScreenPlayerRef.current?.seek(currentPlayTime);
+                            }}
+                            style={styles.fullScreenVideoStyle}
+                        />
+                    </TouchableWithoutFeedback>
+                    {showControls && (
+                        <TouchableOpacity
+                            style={styles.controlBtnContainerStyle}
+                            onPress={() => {
+                                fullScreenHandler(false);
+                            }}
+                        >
+                            <Icon name="fullscreen-exit" color="white" size={22} />
+                        </TouchableOpacity>
+                    )}
+                    {status === StatusEnum.LOADING && (
+                        <ActivityIndicator
+                            color={styles.indicator.color}
+                            size="large"
+                            style={styles.fullScreenActivityIndicatorStyle}
+                        />
+                    )}
+                    {status === StatusEnum.ERROR && (
+                        <Text style={styles.errorMessage}>The video failed to load :(</Text>
+                    )}
+                </Modal>
+            )}
+            <View style={styles.container}>
+                {status === StatusEnum.LOADING && <ActivityIndicator color={styles.indicator.color} size="large" />}
+                {status === StatusEnum.ERROR && <Text style={styles.errorMessage}>The video failed to load :(</Text>}
+                {!fullScreen && (
+                    <TouchableWithoutFeedback style={styles.container} onPress={() => handleShowControls()}>
+                        <Video
+                            {...videoProps}
+                            paused={!props.autoStart && !currentPlayTime}
+                            onLoad={data => {
+                                setStatus(StatusEnum.READY);
+                                setVideoAspectRatio(data.naturalSize.width / data.naturalSize.height);
+                                playerRef.current?.seek(currentPlayTime);
+                            }}
+                            ref={playerRef}
+                            style={status !== StatusEnum.READY ? { height: 0 } : styles.video}
+                            onProgress={({ currentTime }) => {
+                                setCurrentPlayTime(currentTime);
+                            }}
+                        />
+                    </TouchableWithoutFeedback>
+                )}
+                {isAndroid && showControls && (
+                    <TouchableOpacity
+                        onPress={() => {
+                            fullScreenHandler(true);
+                        }}
+                        style={styles.controlBtnContainerStyle}
+                    >
+                        <Icon name="fullscreen" color="white" size={22} />
+                    </TouchableOpacity>
+                )}
+            </View>
+        </Fragment>
     );
 }