+++ packages/pluggableWidgets/video-player-native/CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this widget will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Android full screen support.
 
 ## [3.2.0] - 2022-2-21